+++ tests/testthat/test-ckdepi_2021_egfr.R
@@ -2,7 +2,7 @@
 METHOD = "CKDEPI 2021"
 test_that("ckdepi_2021_egfr works for numerical input", {
   expect_equal(
-    ckdepi_2021_egfr_cystatin(sexf = FALSE, age = 24, creat = 1, cystc = 1) %>%
+    .egfr_ckdepi_2021_cystatin(sexf = FALSE, age = 24, creat = 1, cystc = 1) %>%
       round(3),
     97.570,
     ignore_attr = TRUE
@@ -11,7 +11,7 @@ test_that("ckdepi_2021_egfr works for numerical input", {
 
 test_that("ckdepi_2021_egfr works for vector input", {
   expect_equal(
-    ckdepi_2021_egfr_cystatin(
+    .egfr_ckdepi_2021_cystatin(
       c(FALSE, TRUE, FALSE, TRUE),
       c(24, 24, 23, 24),
       c(1, 1, 2, 1),
@@ -32,7 +32,7 @@ test_that("ckdepi_2021_egfr works for dataframe columns", {
     "CYSTC" = c(0.4, 0.8, 1, 2)
   )
   expect_equal(
-    ckdepi_2021_egfr_cystatin(df$SEXN, df$AGE, df$CREAT, df$CYSTC) %>% round(3),
+    .egfr_ckdepi_2021_cystatin(df$SEXN, df$AGE, df$CREAT, df$CYSTC) %>% round(3),
     c(145.193, 97.491, 67.182, 47.793),
     ignore_attr = TRUE
   )
@@ -48,7 +48,7 @@ test_that("ckdepi_2021_egfr can be used in a mutate", {
   )
   df <- df %>%
     dplyr::mutate(
-      ckdepi_2021_egfr = ckdepi_2021_egfr_cystatin(SEXN, AGE, CREAT, CYSTC)
+      ckdepi_2021_egfr = .egfr_ckdepi_2021_cystatin(SEXN, AGE, CREAT, CYSTC)
     )
 
   expect_equal(
@@ -71,7 +71,7 @@ test_that("ckdepi_2021_egfr can be used within mutate after a group_by", {
   df <- df %>%
     dplyr::group_by(ID) %>%
     dplyr::mutate(
-      ckdepi_2021_egfr = ckdepi_2021_egfr_cystatin(SEXN, AGE, CREAT, CYSTC)
+      ckdepi_2021_egfr = .egfr_ckdepi_2021_cystatin(SEXN, AGE, CREAT, CYSTC)
     )
   expect_equal(
     df$ckdepi_2021_egfr %>% round(3),
@@ -81,7 +81,7 @@ test_that("ckdepi_2021_egfr can be used within mutate after a group_by", {
 })
 
 test_that("ckdepi_2021_egfr_cystatin sets units attribute", {
-  result <- ckdepi_2021_egfr_cystatin(sexf = FALSE, age = 24, creat = 1, cystc = 1)
+  result <- .egfr_ckdepi_2021_cystatin(sexf = FALSE, age = 24, creat = 1, cystc = 1)
   expect_equal(as.character(units(result)), "mL/(min*bsa_ref)")
 })
 
@@ -107,33 +107,33 @@ test_that("ckdepi_2021_egfr won't work for character Sex", {
     df <- df %>%
       dplyr::group_by(ID) %>%
       dplyr::mutate(
-        ckdepi_2021_egfr = ckdepi_2021_egfr_cystatin(SEX, AGE, CREAT, CYSTC) #error here due to non numerical sex
+        ckdepi_2021_egfr = .egfr_ckdepi_2021_cystatin(SEX, AGE, CREAT, CYSTC) #error here due to non numerical sex
       )
   )
 })
 
 test_that("ckdepi_2021_egfr messages about missing values", {
-  expect_message(ckdepi_2021_egfr_cystatin(NA, 24, 1, 0.9), "sexf contains ")
-  expect_message(ckdepi_2021_egfr_cystatin(FALSE, NA, 1, 0.9), "age contains ")
+  expect_message(.egfr_ckdepi_2021_cystatin(NA, 24, 1, 0.9), "sexf contains ")
+  expect_message(.egfr_ckdepi_2021_cystatin(FALSE, NA, 1, 0.9), "age contains ")
   expect_message(
-    ckdepi_2021_egfr_cystatin(FALSE, 24, NA, 0.9),
+    .egfr_ckdepi_2021_cystatin(FALSE, 24, NA, 0.9),
     "creat contains "
   )
-  expect_message(ckdepi_2021_egfr_cystatin(FALSE, 24, 1, NA), "cystc contains ")
+  expect_message(.egfr_ckdepi_2021_cystatin(FALSE, 24, 1, NA), "cystc contains ")
 })
 
 test_that("ckdepi_2021_egfr warns about recycling", {
   sexf <- c(TRUE, FALSE, TRUE)
   expect_warning(
-    ckdepi_2021_egfr_cystatin(sexf, 24, 1, 0.9),
+    .egfr_ckdepi_2021_cystatin(sexf, 24, 1, 0.9),
     "Inputs have different lengths! Please check data."
   )
   expect_warning(
-    ckdepi_2021_egfr_cystatin(TRUE, c(25, 30), 1, 0.9),
+    .egfr_ckdepi_2021_cystatin(TRUE, c(25, 30), 1, 0.9),
     "Inputs have different lengths! Please check data."
   )
   expect_warning(
-    ckdepi_2021_egfr_cystatin(TRUE, 24, c(1.0, 1.2), 0.9),
+    .egfr_ckdepi_2021_cystatin(TRUE, 24, c(1.0, 1.2), 0.9),
     "Inputs have different lengths! Please check data."
   )
 })